tools/requirements.txt
PyGithub>=1.55
pandoc>=2.0
pyyaml>=6.0